requirements.txt
requests
python-dotenv
flask
flask-wtf
urllib3==1.26.6